requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
certifi==2021.10.8
charset-normalizer==2.0.12; python_version >= '3'
click==8.1.3; python_version >= '3.7'
flask-cors==3.0.10
flask==2.1.2
idna==3.3; python_version >= '3'
importlib-metadata==4.11.3; python_version < '3.10'
itsdangerous==2.1.2; python_version >= '3.7'
jinja2==3.1.2; python_version >= '3.7'
markupsafe==2.1.1; python_version >= '3.7'
pymongo==4.1.1
python-dotenv==0.20.0
requests==2.27.1
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
urllib3==1.26.9; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
uwsgi==2.0.20
werkzeug==2.1.2; python_version >= '3.7'
zipp==3.8.0; python_version >= '3.7'
